refactor(list,show): gather the filter-consuming modes where the guard can name them
Pure moves, no logic changed. The recursive --parent tree walk, the
--watch poll loaders, the proxied route's builder call and --current on
both routes move into cmd/bd/list_show_filter_modes.go.

They move because they are the answer to "what in these commands still
legitimately needs the FILTER as a value", and the next commit turns the
rest of those files into a place where naming one is a lint failure. A
mode that consumes the filter now lives in a file the config names, so
widening that hole is an edit a reviewer sees rather than a line added
anywhere in cmd/bd.
