+++ cmd/bd/list.go
@@ -4,10 +4,7 @@ import (
 	"context"
 	"fmt"
 	"os"
-	"os/signal"
 	"strings"
-	"syscall"
-	"time"
 
 	"github.com/spf13/cobra"
 	"github.com/steveyegge/beads/internal/metrics"
@@ -35,180 +32,6 @@ func withStorage(ctx context.Context, store storage.DoltStorage, dbPath string, 
 		return fn(roStore)
 	}
 	return fmt.Errorf("no storage available")
-}
-
-// getHierarchicalChildren handles the --tree --parent combination logic.
-// baseFilter carries CLI filters (--type, --status, etc.) through the recursive walk.
-func getHierarchicalChildren(ctx context.Context, store storage.DoltStorage, dbPath string, parentID string, baseFilter types.IssueFilter) ([]*types.Issue, error) {
-	// First verify that the parent issue exists
-	var parentIssue *types.Issue
-	err := withStorage(ctx, store, dbPath, func(s storage.DoltStorage) error {
-		var err error
-		parentIssue, err = s.GetIssue(ctx, parentID)
-		return err
-	})
-	if err != nil {
-		return nil, fmt.Errorf("error checking parent issue: %v", err)
-	}
-	if parentIssue == nil {
-		return nil, fmt.Errorf("parent issue '%s' not found", parentID)
-	}
-
-	// Use recursive search to find all descendants using the same logic as --parent filter.
-	// The parent itself is NOT included in the result set — only actual children and
-	// their descendants. This matches the behavior of --json and --flat (GH#3349).
-	allDescendants := make(map[string]*types.Issue)
-
-	err = findAllDescendants(ctx, store, dbPath, parentID, baseFilter, allDescendants)
-	if err != nil {
-		return nil, fmt.Errorf("error finding descendants: %v", err)
-	}
-
-	if len(allDescendants) == 0 {
-		return nil, nil
-	}
-
-	// Include the parent as the tree root only when descendants exist,
-	// so the tree renderer can draw the hierarchy with the parent at the top.
-	allDescendants[parentID] = parentIssue
-
-	treeIssues := make([]*types.Issue, 0, len(allDescendants))
-	for _, issue := range allDescendants {
-		treeIssues = append(treeIssues, issue)
-	}
-
-	return treeIssues, nil
-}
-
-// findAllDescendants recursively finds all descendants using parent filtering.
-// baseFilter carries CLI filters (--type, --status, etc.) so the tree respects them.
-func findAllDescendants(ctx context.Context, store storage.DoltStorage, dbPath string, parentID string, baseFilter types.IssueFilter, result map[string]*types.Issue) error {
-	var children []*types.Issue
-	err := withStorage(ctx, store, dbPath, func(s storage.DoltStorage) error {
-		filter := baseFilter
-		filter.ParentID = &parentID
-		filter.Limit = 0 // unlimited per level to avoid truncating the tree walk
-		var err error
-		children, err = s.SearchIssues(ctx, "", filter)
-		return err
-	})
-	if err != nil {
-		return err
-	}
-
-	for _, child := range children {
-		if _, exists := result[child.ID]; !exists {
-			result[child.ID] = child
-			err = findAllDescendants(ctx, store, dbPath, child.ID, baseFilter, result)
-			if err != nil {
-				return err
-			}
-		}
-	}
-
-	return nil
-}
-
-// watchIssues polls for changes and re-displays (GH#654)
-// Uses polling instead of fsnotify because Dolt stores data in a server-side
-// database, not files — file watchers never fire.
-type watchListDependencyStore interface {
-	GetAllDependencyRecords(ctx context.Context) (map[string][]*types.Dependency, error)
-}
-
-func loadWatchedIssues(ctx context.Context, store storage.DoltStorage, filter types.IssueFilter, ready bool, parentID string, sortBy string, reverse bool) ([]*types.Issue, error) {
-	if ready {
-		issues, err := store.GetReadyWork(ctx, workapi.ReadyFilterFromIssueFilter(workapi.WithFetchOneExtra(filter)))
-		if err != nil {
-			return nil, err
-		}
-		workapi.SortIssues(issues, sortBy, reverse)
-		return issues, nil
-	}
-
-	if parentID != "" {
-		issues, err := getHierarchicalChildren(ctx, store, "", parentID, filter)
-		if err != nil {
-			return nil, err
-		}
-		// getHierarchicalChildren builds its result from a map, so normalize the
-		// slice before snapshot comparison to avoid spurious redraws.
-		workapi.SortIssues(issues, "id", false)
-		return issues, nil
-	}
-
-	issues, err := store.SearchIssues(ctx, "", workapi.WithFetchOneExtra(filter))
-	if err != nil {
-		return nil, err
-	}
-	workapi.SortIssues(issues, sortBy, reverse)
-	return issues, nil
-}
-
-func displayWatchedIssueList(ctx context.Context, store watchListDependencyStore, issues []*types.Issue) {
-	var allDeps map[string][]*types.Dependency
-	if store != nil {
-		deps, err := store.GetAllDependencyRecords(ctx)
-		if err == nil {
-			allDeps = deps
-		}
-	}
-	displayPrettyListWithDeps(issues, true, allDeps)
-}
-
-// watchIssues returns an error only for the initial query — a failure there
-// means bd list --watch never displayed anything, so (unlike a mid-poll
-// refresh failure, which just logs and keeps the last good snapshot on
-// screen) it must propagate to the caller. In particular this lets
-// runListCore route a MaxRows cap violation through handleMaxRowsError for
-// exit-code-2 semantics instead of watchIssues swallowing it and exiting 0
-// (be-x42v.4 follow-up, review MUST-FIX 5).
-func watchIssues(ctx context.Context, store storage.DoltStorage, filter types.IssueFilter, ready bool, parentID string, sortBy string, reverse bool, effectiveLimit int) error {
-	// Initial display
-	issues, err := loadWatchedIssues(ctx, store, filter, ready, parentID, sortBy, reverse)
-	if err != nil {
-		return err
-	}
-	// The order is already loadWatchedIssues', which the snapshot comparison
-	// below depends on; what is left is the cut and its verdict, and those are
-	// the shared epilogue's on every other listing this command has.
-	issues, truncated := workapi.FinishPage(issues, "", false, effectiveLimit, false)
-	displayWatchedIssueList(ctx, store, issues)
-	printTruncationHint(truncated, effectiveLimit)
-	lastSnapshot := issueSnapshot(issues)
-
-	fmt.Fprintf(os.Stderr, "\nWatching for changes... (Press Ctrl+C to exit)\n")
-
-	// Handle Ctrl+C — deferred Stop prevents signal handler leak
-	sigChan := make(chan os.Signal, 1)
-	signal.Notify(sigChan, os.Interrupt, syscall.SIGTERM)
-	defer signal.Stop(sigChan)
-
-	pollInterval := 2 * time.Second
-	ticker := time.NewTicker(pollInterval)
-	defer ticker.Stop()
-
-	for {
-		select {
-		case <-sigChan:
-			fmt.Fprintf(os.Stderr, "\nStopped watching.\n")
-			return nil
-		case <-ticker.C:
-			issues, err := loadWatchedIssues(ctx, store, filter, ready, parentID, sortBy, reverse)
-			if err != nil {
-				fmt.Fprintf(os.Stderr, "Error refreshing issues: %v\n", err)
-				continue
-			}
-			issues, truncated := workapi.FinishPage(issues, "", false, effectiveLimit, false)
-			snap := issueSnapshot(issues)
-			if snap != lastSnapshot {
-				lastSnapshot = snap
-				displayWatchedIssueList(ctx, store, issues)
-				printTruncationHint(truncated, effectiveLimit)
-				fmt.Fprintf(os.Stderr, "\nWatching for changes... (Press Ctrl+C to exit)\n")
-			}
-		}
-	}
 }
 
 // issueSnapshot builds a comparable string from issue IDs, statuses, and
